SUP-1050, Hitting back button after any my-dashboard link breaks my-dashboard
It was happening because getHandle method is not being called on back button which was happening because of page being marked as cached. For now, workaround it by marking the page non cached because down v2 of dashboard will be a separate app which would not be depdendent on tc-site.