Small modification of condition to show the Challenge Docs download link
Now the link should be visible for all active challenge, no matter what
phase they are in. Note that there are additional coniditions added in
the HTML markup here: src/js/app/challenge-details/partials/ng-content-
challenge-downloads.tpl.html which also demand the user to be logged in
and registered (different messages are shown instead of the link, if it
is not true).